moves controller
